Move some crypto utilities in libs/crypto
