+++ cmd/vega/genesis/generate.go
@@ -9,8 +9,8 @@ import (
 
 	"code.vegaprotocol.io/vega/assets"
 	"code.vegaprotocol.io/vega/config"
-	"code.vegaprotocol.io/vega/crypto"
 	"code.vegaprotocol.io/vega/genesis"
+	"code.vegaprotocol.io/vega/libs/crypto"
 	"code.vegaprotocol.io/vega/logging"
 	"code.vegaprotocol.io/vega/netparams"
 	"code.vegaprotocol.io/vega/nodewallet"
