+++ execution/market.go
@@ -12,9 +12,9 @@ import (
 	"time"
 
 	"code.vegaprotocol.io/vega/collateral"
-	"code.vegaprotocol.io/vega/crypto"
 	"code.vegaprotocol.io/vega/events"
 	"code.vegaprotocol.io/vega/fee"
+	"code.vegaprotocol.io/vega/libs/crypto"
 	"code.vegaprotocol.io/vega/liquidity"
 	liquiditytarget "code.vegaprotocol.io/vega/liquidity/target"
 	"code.vegaprotocol.io/vega/logging"
