+++ libs/crypto/hash.go
