+++ nodewallet/nodewallet_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"code.vegaprotocol.io/vega/config/encoding"
-	"code.vegaprotocol.io/vega/crypto"
+	"code.vegaprotocol.io/vega/libs/crypto"
 	"code.vegaprotocol.io/vega/logging"
 	"code.vegaprotocol.io/vega/nodewallet"
 	"code.vegaprotocol.io/vega/nodewallet/eth/mocks"
